disable farp plugin with module conf and remove manual load list
Due to an error in the settings, the manual load list of plugins
for the charon daemon was not used by the software.
To be useful, `load_modular = yes` would have to be set.

Because of errors in module loading and discourage of using the
manual load list by the strongswan documentation, the modules will
now be configured by their respective configuration files.